+++ CHANGELOG.md
@@ -33,6 +33,7 @@
   - Remove sdx-common logging
   - Add validation on tx_id search to force a valid UUID string to be entered
   - FTP files write to memory instead of writing temporary files
+  - Add pagination in store data
 
 ### 1.7.0 2017-11-30
   - Deploying new console